requirements.txt
streamlit
aiogram
pandas
openpyxl
langchain
langchain-openai
langchain-community
qdrant-client
python-dotenv
openai
tiktoken
httpx
watchdog
